requirements/requirement.txt
# pip install -r requirements/requirement.txt

Django==3.0.7
djangorestframework==3.11.0
django-cors-headers==3.4.0
django-filter==2.3.0

asgiref==3.2.10
pytz==2020.1
sqlparse==0.3.1

black==19.10b0
